Add a palette selector & make the context selector slightly smaller.
Instead of displaying all tools, the context selector now only displays
the rectangular selection, eraser, color picker, and the last tool used
that is not one of those.

The palette set was copied from the 100 Palette Challenge:

http://fallenzephyr.deviantart.com/art/100-Palette-Challenge-462719186

See `selector.coffee` for a short explanation of the contents
of `palettes.dat`.